Consolidating the Algolia services class into the datasource class.
